Package webmacs in local overlay
Not entirely without hacks...

Since the dateparser overlay doesn't seem to work correctly currently,
I had to duplicate the fix in here. This hack has to be removed when
the issue with dateparser gets resolved upstream.

"export CC=g++": since CC is defined inside the stdenv(?) we need to
replicate the hack inside webmacs' setup.py in preBuild.