overlays/local/pkgs: Remove dependency on networkmanager in libproxy
This thing is depended on by glib-networking, which in turn is
depended on by a lot of stuff (like gimp, apparently... modern
software doesn't even make sense)

I did this because I encountered a cache miss on network manager, and
I didn't feel like building things I don't use.

OTOH this might cause a cascade of cache misses due to being so
ubiquitous, I'll just have to find out if it's worth it or not.