overlays/local/pkgs: package easystroke
Take some patches from gentoo (not all are neccessary since we are
using git master).

The easystroke-0.6.0-gentoo.patch is currently disabled as it needs to
be rewritten for git master. The patch isn't strictly neccessary, but
it might be useful to make the makefile more well-behaved (e.g. now it
strips the binaries and such)