overlays: Specify owner and group for simpleserver security wrapper
As is now needed for security.wrappers:

```
The `security.wrappers` option now requires to always specify an owner, group and whether the setuid/setgid bit should be set.
  This is motivated by the fact that before NixOS 21.11, specifying either setuid or setgid but not owner/group resulted in wrappers owned by nobody/nogroup, which is unsafe.
```

In my testing, however, I found that it didn't actually complain about
not explicitly stating the setgid and setuid options, and only setting
owner and group was enough. I did specify both of these as well though
for good measure (maybe the checks in nixpkgs for this will work
properly in the future?)