bump wip-pinebook-pro + changes needed to deal with that
Perhaps now I don't need to build the entire kernel

Also includes the changes needed in my config to deal with changes in
wip-pinebook-pro. Note that I do not think that it can install u-boot
into the eMMC itself any longer, but it will be good enough for now to
just make sure that it creates the entries in extlinux.conf. Later we
will likely want to be switching to Tow-Boot on the SPI anyway (but
probably continue extlinux.conf-based booting. I do not see the point
of EFI-booting really)