machines/nazrin: back to latest greatest kernel
Now that we are using a USB wifi thing with the r8118eu driver, which
just hit staging in the kernel in 5.15, I suspect we will want as new
of a kernel as possible with whatever fixes etc.

While broadcom-sta compiles in 5.15, it seems to be completely dead.
On 5.17 it won't even compile, so we disable the broadcom-sta driver
in this commit.

TODO: get a USB thing with better linux driver (and maybe 802.11ac?)