.config/nixpkgs/usbee: Get rid of most GL wrapped things
Seems to work fine without this these days? We'll try it at least.
This commit can be reverted if need be.

We still keep the pulseview one around because it also does some
python shenanigans.

Also the nixGLIntel and such wrappers I keep around a little bit
longer just in case.