home/desktop.nix: stash away experiment in adding fcitx5 with home-manager
Not sure if this makes sense to do, adding it in the normal nixpkgs
way seems to work just as well with autostarting it and everything due
to xdg-autostart shenanigans (at least when running a full-blown DE
like I am testing out right now. Might be nice to configure up my
lightweight alternatives to work similarly though? Or maybe not
xdg-autostart stuff is sometimes annoying)