machines/zeke: Switch to ibus as input method
This is probably the better pick for GNOME, but not sure if it is for
KDE. It seems to work for me now though.

I'm making this change to this machine only and not globally, since I
have some machines that rely on starting the input method manually
with .xsession still.

Eventually I might make this transition for all machines, however.