emacs-parse: support for more special forms
This isn't really neccessary in most common usage, mostly just an
extra flair. I guess it is useful in that we do not try to parse
inside quotes, or the variable name portion of a let or so.

To be really fancy we should make processing more
progn/expression-aware though (that is walk should only handle an
expression, and we'd have another function use map on it)