sumireko: Do not use my rockchip overlay for now
The mpv overlay in there does not build properly. Same sort of problem
as with my custom mpv build in the regular overlay :<. (I need to
figure out how to defeat the wrapper there).

On another hand, I am also not sure whether this special version of
ffmpeg is needed any longer to get hwdec with mpv on this thingy. I
think I need to do quite a bit more testing (perhaps need to test if I
can get this working on an arch install or something as well, I don't
remember having much luck with this particular thing on nixos for the
time being...)