home/home.*.nix: Hacky way to ensure XDG_CURRENT_DESKTOP is set for ratpoison
This might be useful for desktop portal configs and what
not. (set up using xdg.portal.config or something)

TODO: Maybe it'd be better to put this inside my ratpoison config
itself? This sort of thing makes sense to have on non-nixos-systems as
well (although I'd need to figure out how to make it conditional when
there isn't systemd or dbus and stuff I guess. or just live with some
error message in the ratpoison log)