overlays/local: bump wayvr-dashboard version to 0.2.8
I think this is supposed to be 0.2.8 at least. But it seems like they
have forgotten to tag this commit (latest tag at time of writing is 0.2.7)

Also updating the default-fromlemon.nix file even though I'm not using
it. Although I guess that one might need to bum pthe npmDepsHash as
well or so?