Try and make an emacs that is wrapped with language servers for eglot
Is this the way to do it. Or perhaps it's better to just stick stuff
into environment.systemPackages and do things in a more traditional,
less nixy manner...

One downside to wrapping like this is that by default we weren't
getting emacsclient and etags etc. included. We have to explicitly
name them on the pkgs.runCommand invocation. Maybe there is a better
way of doing the wrapping that'll automatically pick it up though?

With this update we stop using the home manager program.emacs and
service.emacs modules, because it is too fancy to make this wrapper
stuff work. Anyway, the systemd service for emacs is borked with
wayland sessions for some reason (has to be manually restarted...).
Might as well just start emacs manually then. Maybe services.emacs
could be made to work somehow if it took an arbitrary emacs package
input. I dunno.