overlays/local/pkgs: Re-enable the ALCOM pkg using a pinned version of nixos_unstable
Doing a bit of an ugly?-hack this way. Yay. Who needs flakes anyway?

Maybe we could also use this for things like my custom backporting of
lact? Although my current way of doing that one has the benefit of
not unneccessarily pulling in nixos-unstable dependencies, since the
version-bumped lact seems to build just fine on top of nixos-25.05 deps.