machines/{leon,nazrin}: Add support for atari AHDI partition tables
Because I want to be able to handle HD images that were formatted in
this manner.

Unfortunately this means we are going to have to rebuild the kernel,
as this'll cause a cache miss.

For nazrin it is simply a case of re-enabling what I had previously
commented out (probably because I was trying to avoid having to
recompile the kernel at some point).