leon/xr: note in comments about settings to service.monado I want to test in the future
This might make it so we don't have to add as much to commandline in
steam.

The monado environment stuff might perhaps also make the
wivrn-dashboard-trackers wrapper script unneccessary? OTOH it's
perhaps nice to be able to choose on the fly whether we want to scan
for steamvr_lh trackers or not